// shared/api/generated/model/facilityItemInfraType.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Dingle Swagger API
 * Dingle API
 * OpenAPI spec version: v1.0.0
 */

export type FacilityItemInfraType = typeof FacilityItemInfraType[keyof typeof FacilityItemInfraType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FacilityItemInfraType = {
  CCTV: 'CCTV',
  POLICE_STATION: 'POLICE_STATION',
  EARTHQUAKE_SHELTER: 'EARTHQUAKE_SHELTER',
  CIVIL_SHELTER: 'CIVIL_SHELTER',
  CONVENIENCE_STORE: 'CONVENIENCE_STORE',
  MARKET: 'MARKET',
  HOSPITAL: 'HOSPITAL',
  EMERGENCY_MEDICAL_CENTER: 'EMERGENCY_MEDICAL_CENTER',
  HEAT_SHELTER: 'HEAT_SHELTER',
  COLD_SHELTER: 'COLD_SHELTER',
  SAFETY_LIGHT: 'SAFETY_LIGHT',
} as const;